chore(packages): consolidate middleware integ tests
