+++ packages/middleware-apply-body-checksum/package.json
@@ -9,8 +9,7 @@
     "build:types": "tsc -p tsconfig.types.json",
     "build:types:downlevel": "downlevel-dts dist-types dist-types/ts3.4",
     "clean": "rimraf ./dist-* && rimraf *.tsbuildinfo",
-    "test": "jest --coverage",
-    "test:integration": "jest --config jest.config.integ.js"
+    "test": "jest --coverage"
   },
   "main": "./dist-cjs/index.js",
   "module": "./dist-es/index.js",
