--- packages/middleware-retry/jest.config.integ.js
@@ -1,4 +1,0 @@
-module.exports = {
-  preset: "ts-jest",
-  testMatch: ["**/*.integ.spec.ts"],
-};
