+++ private/aws-middleware-test/jest.config.js
@@ -1,4 +1,4 @@
 module.exports = {
   preset: "ts-jest",
-  testMatch: ["**/*.integ.spec.ts"],
+  testMatch: ["**/*.spec.ts"],
 };
